Rename LandingLine to AssignedLine; add ReleaseLines stage
Renamed the property `LandingLine` to `AssignedLine` in the `Airplane` class and updated all references accordingly. Added a new method `IsDepartured` to the `Airplane` class to check if an airplane has departed. Introduced a new simulation stage `ReleaseLines` in the `AirportSimulationBuilder` class to release lines assigned to departed airplanes. Created a new file `ReleaseLines.cs` to define the `ReleaseLines` class, implementing the `ISimulationStage<AirportEnvironment>` interface.