Refactor to remove generic type parameter from interfaces
The changes in the code primarily involve the removal of the generic type parameter `LifeMatrixEnvironment` from the `ISimulationContext` and `ISimulationStage` interfaces in several classes. This change simplifies the method signatures and makes the code more flexible by allowing the `ISimulationContext` to work with different types of simulation environments.

In `GameOfLifeDrawer.cs`, the `Draw` method has been updated to remove the generic type parameter from `ISimulationContext`. The method now retrieves the simulation environment using the `GetSimulationEnvironment<LifeMatrixEnvironment>` method. Additionally, the code has been refactored to use the `environment` variable instead of directly accessing `context.SimulationEnvironment`.

In `InitializeSize.cs`, the `InitializeSize` class has been updated to remove the generic type parameter from the `ISimulationStage` interface. The `Execute` method now retrieves the simulation environment using the `GetSimulationEnvironment<LifeMatrixEnvironment>` method and initializes it with the specified width and height.

In `UpdateMatrix.cs`, the `UpdateMatrix` class has been updated similarly to remove the generic type parameter from the `ISimulationStage` interface. The `Execute` method now retrieves the simulation environment using the `GetSimulationEnvironment<LifeMatrixEnvironment>` method and updates it with the specified number of threads.